Herald get the fuck out

+++ app/helpers/application_helper.rb
@@ -23,14 +23,16 @@ module ApplicationHelper
   end
 
   def sharing_btn(user, size)
-    if user == current_user
-      link_to "That's you!", "#", :class => "btn btn-social #{size}", :rel => "tooltip", :data => { :title => "Yeah really..", :placement => "left" }
-    else
-      if current_user.isSharingWith(user)
-        friendship = current_user.friendships.where(:friend_id => user.id).first
-        link_to content_tag("i", "", class: "icon-heart")+" Sharing", friendship_path(friendship), :method => :delete, :remote => true, :class => "btn btn-social btn-unsocial #{size}"
+    if @herald.nil? # if isn't accessed via herald
+      if user == current_user
+        link_to "That's you!", "#", :class => "btn btn-social #{size}", :rel => "tooltip", :data => { :title => "Yeah really..", :placement => "left" }
       else
-        link_to "Share", friendships_path(:friend => user), :method => :post, :remote => true, :class => "btn btn-social #{size}"
+        if current_user.isSharingWith(user)
+          friendship = current_user.friendships.where(:friend_id => user.id).first
+          link_to content_tag("i", "", class: "icon-heart")+" Sharing", friendship_path(friendship), :method => :delete, :remote => true, :class => "btn btn-social btn-unsocial #{size}"
+        else
+          link_to "Share", friendships_path(:friend => user), :method => :post, :remote => true, :class => "btn btn-social #{size}"
+        end
       end
     end
   end
